+++ R/generics.R
@@ -38,6 +38,11 @@ setGeneric("createGiottoPolygon",
             function(x, ...) standardGeneric("createGiottoPolygon"))
 
 
+# Object reconnection ####
+# Some objects may operate base on on-disk files.
+setGeneric("reconnect", function(x, ...) standardGeneric("reconnect"))
+
+
 # Object Characteristics ####
 
 ## nrow() S4 generic ####