+++ R/methods-plot.R
@@ -43,7 +43,6 @@ setMethod("plot", signature(x = "giottoImage", y = "missing"), function(x, y, ..
 #' plot(gimg, max_intensity = 100)
 #' }
 #'
-#'
 #' @export
 setMethod(
     "plot",
@@ -98,7 +97,6 @@ setMethod(
 #' gpoly <- GiottoData::loadSubObjectMini("giottoPolygon")
 #' plot(gpoly)
 #' plot(gpoly, type = "centroid")
-#'
 #'
 #' @export
 setMethod(
@@ -184,7 +182,6 @@ setMethod(
 #' # plot specific feature(s)
 #' plot(gpoints, feats = featIDs(gpoints)[seq_len(4)], raster = FALSE)
 #'
-#'
 #' @export
 setMethod(
     "plot", signature(x = "giottoPoints", y = "missing"),
@@ -206,7 +203,6 @@ setMethod(
 #' sl <- GiottoData::loadSubObjectMini("spatLocsObj")
 #' plot(sl)
 #'
-#'
 #' @export
 setMethod("plot", signature(x = "spatLocsObj", y = "missing"), function(x, ...) {
     if ("sdimz" %in% colnames(x)) {
@@ -225,7 +221,6 @@ setMethod("plot", signature(x = "spatLocsObj", y = "missing"), function(x, ...) 
 #' d <- GiottoData::loadSubObjectMini("dimObj")
 #' plot(d)
 #' plot(d, dims = c(3, 5))
-#'
 #'
 #' @export
 setMethod(
@@ -650,53 +645,52 @@ setMethod("plot", signature(x = "spatialNetworkObj", y = "missing"), function(x,
 #' @keywords internal
 #' @noRd
 .plot_giotto_points_all <- function(
-    x, size = 600, force_size = FALSE, dens = FALSE, col = NULL, background, ...
-) {
-  pargs <- list(...)
-  rargs <- list()
-  e <- ext(x)
-  e_r <- range(e)
+        x, size = 600, force_size = FALSE, dens = FALSE, col = NULL, background, ...) {
+    pargs <- list(...)
+    rargs <- list()
+    e <- ext(x)
+    e_r <- range(e)
 
-  # decide rasterization resolution
-  # Select res that results in a major axis with length equal to size param,
-  # up to a maximum resolution of 1 (1:1 with extent dims),
-  # but with a min dim px of 100 (to help with cases where extent is small)
-  res <- max(e_r / size[1L])
-  if (!isTRUE(force_size)) {
-    res <- max(res, 1)
-    res <- min(res, c(e_r / 100))
-  }
-
-  # rasterization
-  r <- terra::rast(e, res = res)
-  if (isTRUE(dens)) rargs$fun <- "count"
-  rargs$y <- r
-  rargs$x <- x[]
-  r2 <- do.call(terra::rasterize, args = rargs)
-
-  # plotting
-  pargs$x <- r2
-  if (is.null(col)) {
-    if (isTRUE(dens)) {
-      pal <- grDevices::hcl.colors(n = 256)
-    } else {
-      pal <- c("black", "white")
+    # decide rasterization resolution
+    # Select res that results in a major axis with length equal to size param,
+    # up to a maximum resolution of 1 (1:1 with extent dims),
+    # but with a min dim px of 100 (to help with cases where extent is small)
+    res <- max(e_r / size[1L])
+    if (!isTRUE(force_size)) {
+        res <- max(res, 1)
+        res <- min(res, c(e_r / 100))
     }
-    pargs$col <- pal[2L:length(pal)]
-    pargs$background <- pal[1L]
-    # replace background col if specifically provided
-    if (!missing(background)) pargs$background <- background
-  } else {
-    if (missing(background)) {
-      pargs$col <- col[2L:length(col)]
-      pargs$background <- col[1L]
-    } else {
-      pargs$col <- col
-      pargs$background <- background
-    }
-  }
 
-  do.call(terra::plot, args = pargs)
+    # rasterization
+    r <- terra::rast(e, res = res)
+    if (isTRUE(dens)) rargs$fun <- "count"
+    rargs$y <- r
+    rargs$x <- x[]
+    r2 <- do.call(terra::rasterize, args = rargs)
+
+    # plotting
+    pargs$x <- r2
+    if (is.null(col)) {
+        if (isTRUE(dens)) {
+            pal <- grDevices::hcl.colors(n = 256)
+        } else {
+            pal <- c("black", "white")
+        }
+        pargs$col <- pal[2L:length(pal)]
+        pargs$background <- pal[1L]
+        # replace background col if specifically provided
+        if (!missing(background)) pargs$background <- background
+    } else {
+        if (missing(background)) {
+            pargs$col <- col[2L:length(col)]
+            pargs$background <- col[1L]
+        } else {
+            pargs$col <- col
+            pargs$background <- background
+        }
+    }
+
+    do.call(terra::plot, args = pargs)
 }
 
 
@@ -810,21 +804,23 @@ setMethod("plot", signature(x = "spatialNetworkObj", y = "missing"), function(x,
 #' @title Plot a giotto polygon object
 #' @param x giottoPolygon object
 #' @param point_size (default = 0.6) size of plotted points when plotting centroids
-#' @param type (default is poly) plot the 'polygon' or its 'centroid'
+#' @param type (default is poly) plot the 'poly' or its 'centroid'
 #' @param ... additional params to pass to plot function
 #' @keywords internal
 #' @noRd
-.plot_giotto_polygon <- function(x, point_size = 0.6,
-    type = c("poly", "centroid"), ...) {
+.plot_giotto_polygon <- function(
+        x, point_size = 0.6,
+        type = c("poly", "centroid"), ...) {
+    a <- list(...)
+
     type <- match.arg(type, choices = c("poly", "centroid"))
-    if (type == "poly") {
-        terra::plot(x = x@spatVector, ...)
-    }
-    if (type == "centroid") {
-        if (!is.null(x@spatVectorCentroids)) {
-            terra::plot(x = x@spatVectorCentroids, cex = point_size, ...)
-        } else {
-            cat("no centroids calculated\n")
+
+    switch(type,
+        "poly" = do.call(terra::plot, args = c(list(x = x@spatVector), a)),
+        "centroid" = {
+            a$cex <- point_size
+            a$x <- centroids(x)
+            do.call(terra::plot, args = a)
         }
-    }
+    )
 }